refactor: manualCount 객체 추출 제거 및 LottoCount안에 원시값으로 합류
