+++ src/main/java/controller/LottoGameApplication.java
@@ -8,8 +8,7 @@ public class LottoGameApplication {
 
     public static void main(String[] args) {
         Money money = inputPurchaseAmountWithValidation();
-        ManualCount manualCount = inputManualCountWithValidation(money);
-        LottoCount lottoCount = new LottoCount(money.getLottoCount(), manualCount.getManualCount());
+        LottoCount lottoCount = inputLottoCountWithValidation(money);
 
         Lottos lottos = LottosGenerator.generateTotal(lottoCount);
         OutputView.printLottos(lottoCount, lottos);
@@ -23,21 +22,21 @@ public class LottoGameApplication {
         OutputView.printProfitRatio(money.calculateProfitRatio(lottoResult));
     }
 
+    private static LottoCount inputLottoCountWithValidation(Money money) {
+        try {
+            return new LottoCount(money.getLottoCount(), InputView.inputManualCount());
+        } catch (IllegalArgumentException | NullPointerException e) {
+            OutputView.printExceptionMessage(e);
+            return inputLottoCountWithValidation(money);
+        }
+    }
+
     private static Money inputPurchaseAmountWithValidation() {
         try {
             return new Money(InputView.inputPurchaseAmount());
         } catch (IllegalArgumentException | NullPointerException e) {
             OutputView.printExceptionMessage(e);
             return inputPurchaseAmountWithValidation();
-        }
-    }
-
-    private static ManualCount inputManualCountWithValidation(final Money money) {
-        try {
-            return new ManualCount(InputView.inputManualCount(), money.getLottoCount());
-        } catch (IllegalArgumentException e) {
-            OutputView.printExceptionMessage(e);
-            return inputManualCountWithValidation(money);
         }
     }
